server/requirements.txt
openai
flask
requests
pandas
openai
python-dotenv
